pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scholar-env"
version = "0.4.0"
description = "OpenEnv environment for scholarly integrity verification"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Nensi Pansuriya"},
    {name = "Krushna Parmar"},
    {name = "Ishita Bhojani"}
]
keywords = ["openenv", "nlp", "research-integrity", "peer-review", "rl-environment"]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "openenv-core>=0.2.0",
    "thefuzz>=0.22.0",
]

[project.scripts]
server = "server.app:main"

[project.urls]
"Homepage" = "https://huggingface.co/spaces/flyingmaverick/scholar-env"

[tool.setuptools]
py-modules = ["corpus", "models", "inference"]
packages = ["server", "server.graders", "scripts", "tests"]
